feat: close issues 218 288 289 290
